add UserController and handle error code 0
