+++ app/config/services.yaml
@@ -23,14 +23,6 @@ services:
 
     # add more service definitions when explicit configuration is needed
     # please note that last definitions always *replace* previous ones
-
-    jms_serializer.object_constructor:
-        alias: jms_serializer.existing_object_constructor
-
-    jms_serializer.existing_object_constructor:
-        class: App\Service\ExistingObjectConstructor
-        arguments: ["@jms_serializer.unserialize_object_constructor"]
-
     user.rel_provider:
         class: App\Service\SubUserRelProvider
         public: true
