+++ app/src/Controller/SecurityController.php
@@ -4,7 +4,6 @@
 namespace App\Controller;
 
 
-use App\Entity\User;
 use OpenApi\Annotations as OA;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -18,8 +17,7 @@ use Symfony\Component\Routing\Annotation\Route;
 #[Route('/', name: "api")]
 class SecurityController extends AbstractController
 {
-    //TODO: add token to header
-//TODO: when subUser gives info check if part of subUser table -> getUser and connect with him
+
     /**
      * Login to gain access to API
      * @OA\Tag(name="Login")
@@ -52,7 +50,6 @@ class SecurityController extends AbstractController
     public function login(Request $request): JsonResponse
     {
         $user = $this->getUser();
-        //subuser credentials
 
         return $this->json([
             'username' => $user->getUserIdentifier(),
@@ -60,27 +57,5 @@ class SecurityController extends AbstractController
         ]);
     }
 
-
-    /**
-     * TODO: Does not work ask if possible ?
-     */
-    #[Route('sub/login', name: "_sub_login", methods: ['POST'])]
-    public function getAccess(Request $request, SubUserRepository $repository, UserRepository $userRepository)
-    {
-        dd($request);
-        /** @var SubUser $subUser */
-        $subUserData= $this->serializer->deserialize($request->getContent(), SubUser::class, 'json');
-        $subUser =  $repository->findOneBy(["username" => $subUserData->getUsername()]);
-
-        $user = $userRepository->find($subUser->getUser()->getId());
-        $user->setRoles(['ROLE_USER']);
-
-        $data = $this->json([
-            "username" => $user->getUserIdentifier(),
-            "password" => $user->getPassword(),
-            "roles" => $user->getRoles()
-        ]);
-
-    }
 
 }
